chore(list): moved list item wrapper outside of list item
