+++ src/components/List/ListItem/ListItem.jsx
@@ -5,140 +5,14 @@ import { Draggable16, ChevronUp16, ChevronDown16 } from '@carbon/icons-react';
 import PropTypes from 'prop-types';
 import isEmpty from 'lodash/isEmpty';
 
-import { EditingStyle, DropLocation } from '../../../utils/DragAndDropUtils';
+import { EditingStyle } from '../../../utils/DragAndDropUtils';
 import { settings } from '../../../constants/Settings';
 
-import ListTarget, { TargetSize } from './ListTarget';
+import ListItemWrapper from './ListItemWrapper';
 
 const { iotPrefix } = settings;
 
 export const ItemType = 'ListItem';
-
-const ListItemWrapper = ({
-  id,
-  editingStyle,
-  expanded,
-  isSelectable,
-  itemWillMove,
-  onItemMoved,
-  onSelect,
-  selected,
-  isDragging,
-  isLargeRow,
-  children,
-  connectDragSource,
-}) => {
-  const body = isSelectable ? (
-    <div
-      role="button"
-      tabIndex={0}
-      className={classnames(
-        `${iotPrefix}--list-item`,
-        `${iotPrefix}--list-item__selectable`,
-        { [`${iotPrefix}--list-item__selected`]: editingStyle ? false : selected },
-        { [`${iotPrefix}--list-item__large`]: isLargeRow },
-        { [`${iotPrefix}--list-item-editable`]: editingStyle }
-      )}
-      data_testid={selected ? 'list-item__selected' : null}
-      onKeyPress={({ key }) => key === 'Enter' && onSelect(id)}
-      onClick={() => {
-        onSelect(id);
-      }}
-    >
-      {children}
-    </div>
-  ) : (
-    <div
-      className={classnames(`${iotPrefix}--list-item`, {
-        [`${iotPrefix}--list-item__large`]: isLargeRow,
-        [`${iotPrefix}--list-item-editable`]: editingStyle,
-      })}
-    >
-      {children}
-    </div>
-  );
-
-  if (editingStyle) {
-    const canNest =
-      editingStyle === EditingStyle.SingleNesting || editingStyle === EditingStyle.MultipleNesting;
-
-    return (
-      <div
-        role="listitem"
-        className={classnames(`${iotPrefix}--list-item-editable--drag-container`, {
-          [`${iotPrefix}--list-item-editable-dragging`]: isDragging,
-        })}
-        ref={instance => {
-          if (connectDragSource) {
-            connectDragSource(instance);
-          }
-        }}
-      >
-        <div
-          className={classnames(`${iotPrefix}--list-item-editable--drop-targets`, {
-            [`${iotPrefix}--list-item__large`]: isLargeRow,
-          })}
-        >
-          {// Renders Nested location only if nesting is allowed
-
-          canNest ? (
-            <ListTarget
-              id={id}
-              isDragging={isDragging}
-              targetPosition={DropLocation.Nested}
-              targetSize={TargetSize.Full}
-              itemWillMove={itemWillMove}
-              onItemMoved={onItemMoved}
-            />
-          ) : null}
-
-          <ListTarget
-            id={id}
-            isDragging={isDragging}
-            targetPosition={DropLocation.Above}
-            itemWillMove={itemWillMove}
-            targetSize={canNest ? TargetSize.Third : TargetSize.Half}
-            onItemMoved={onItemMoved}
-          />
-
-          <ListTarget
-            id={id}
-            isDragging={isDragging}
-            targetPosition={DropLocation.Below}
-            itemWillMove={itemWillMove}
-            targetOverride={expanded ? DropLocation.Nested : null} // If item is expanded then the bottom target will nest
-            targetSize={canNest ? TargetSize.Third : TargetSize.Half}
-            onItemMoved={onItemMoved}
-          />
-        </div>
-
-        {body}
-      </div>
-    );
-  }
-
-  return body;
-};
-
-const ListItemWrapperProps = {
-  dragPreviewText: PropTypes.string,
-  id: PropTypes.string.isRequired,
-  editingStyle: PropTypes.oneOf([
-    EditingStyle.Single,
-    EditingStyle.Multiple,
-    EditingStyle.SingleNesting,
-    EditingStyle.MultipleNesting,
-  ]),
-  expanded: PropTypes.bool.isRequired,
-  isLargeRow: PropTypes.bool.isRequired,
-  isSelectable: PropTypes.bool.isRequired,
-  isDragging: PropTypes.bool.isRequired,
-  onSelect: PropTypes.func.isRequired,
-  selected: PropTypes.bool.isRequired,
-  children: PropTypes.node.isRequired,
-  onItemMoved: PropTypes.func.isRequired,
-  itemWillMove: PropTypes.func.isRequired,
-};
 
 const ListItemPropTypes = {
   id: PropTypes.string.isRequired,
@@ -426,7 +300,6 @@ const ds = DragSource(ItemType, cardSource, (connect, monitor) => ({
   isDragging: monitor.isDragging(),
 }));
 
-ListItemWrapper.propTypes = ListItemWrapperProps;
 ListItem.propTypes = ListItemPropTypes;
 ListItem.defaultProps = ListItemDefaultProps;
